Select wallpaper based on resolution.
I want to use custom wallpaper images for each
machine without hardcoding image paths in the
config. To accomplish this, I created a script
that selects an image from a set directory, so all
I need to do is drop an image in the appropriate
directory, and it will be picked up automatically.